add ResourceRestoreSystems

+++ dungeon/src/contrib/components/EnergyComponent.java
@@ -4,48 +4,48 @@ import core.Component;
 
 public class EnergyComponent implements Component {
 
-  private int maxAmount;
-  private int currentAmount;
+  private float maxAmount;
+  private float currentAmount;
   private float restorePerSecond;
 
-  public EnergyComponent(int maxAmount, int currentAmount, float restorePerSecond) {
+  public EnergyComponent(float maxAmount, float currentAmount, float restorePerSecond) {
     this.maxAmount = maxAmount;
     this.currentAmount = Math.min(currentAmount, maxAmount);
-    this.restorePerSecond=restorePerSecond;
+    this.restorePerSecond = restorePerSecond;
   }
 
   // Getters
-  public int getMaxAmount() {
+  public float getMaxAmount() {
     return maxAmount;
   }
 
-  public int getCurrentAmount() {
+  public float getCurrentAmount() {
     return currentAmount;
   }
 
   // Setters
-  public void setMaxAmount(int maxAmount) {
+  public void setMaxAmount(float maxAmount) {
     this.maxAmount = maxAmount;
     if (currentAmount > maxAmount) {
       currentAmount = maxAmount;
     }
   }
 
-  public void setCurrentAmount(int currentAmount) {
+  public void setCurrentAmount(float currentAmount) {
     this.currentAmount = Math.min(currentAmount, maxAmount);
   }
 
   // Increase/decrease max energy
-  public void increaseMaxAmount(int amount) {
+  public void increaseMaxAmount(float amount) {
     setMaxAmount(maxAmount + amount);
   }
 
-  public void decreaseMaxAmount(int amount) {
+  public void decreaseMaxAmount(float amount) {
     setMaxAmount(Math.max(0, maxAmount - amount));
   }
 
   // Consume/restore current energy
-  public boolean consume(int amount) {
+  public boolean consume(float amount) {
     if (amount <= currentAmount) {
       currentAmount -= amount;
       return true;
@@ -53,13 +53,14 @@ public class EnergyComponent implements Component {
     return false;
   }
 
-  public void restore(int amount) {
+  public void restore(float amount) {
     currentAmount = Math.min(currentAmount + amount, maxAmount);
   }
 
   public float getRestorePerSecond() {
     return restorePerSecond;
   }
+
   public void setRestorePerSecond(float restorePerSecond) {
     this.restorePerSecond = restorePerSecond;
   }
